backend/requirements.txt
# BEACON Backend Requirements
# Install: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# HTTP Client
requests==2.31.0

# LangChain & LangGraph for Agents
langgraph==0.2.0
langchain-core==0.3.0
langchain-valyu==0.1.0

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
